feat: updated TypeDoc to 0.27
The update implies that this module is now ESM only, as nweer versions of TypeDoc promote this. This
involved making changes to the plugins used by theme, the build process, and also updat some
elements to match the newer contents of the default theme. Additionally, the dependency on
gobstones-scripts have been removed, to avoid a circular dependency that complicates the deployment
process. Overall this are the changes:
* Removed the dependency on gobstones-scripts
* Updated the peer dependency of TypeDoc to 0.27.7
* Changed the build process from using tsc only to use rollup
* Updated the output module to be ESM only
* Updated the package-scripts to run on local deffinitions only
* Added the license header locally so it can be used without scripts
* Update the theme components to match the latest version of the default theme
* Updated some CSS styles so that:
    - Now inherited members are shown with an arrow
    - internal, private and protected members are shown with different colors on the listings
* Updated some JS of the theme such that:
    - Now the theme's JS runs after the default theme has completed updating the navbar
    - Fix an issue where the current page is not selected on the left sidebar on some scenarios

BREAKING CHANGE: The peer dependency has ben updated to TypeDoc 0.27.7, so you should update your
typedoc version or this version of the theme will not work.